+++ DTCD-WorkspaceSystem/src/TabsPanelComponent.vue
@@ -25,7 +25,7 @@
         class="TabBtnsList"
       >
         <div
-          v-for="({id, name, isActive, editName}, index) in tabList"
+          v-for="({id, name, isActive, editName, hide}, index) in tabList"
           :key="id"
           :data-tab-id="id"
           class="TabBtn"
@@ -34,7 +34,8 @@
             status_editOn: edit,
             status_editName: editName
           }"
-          @click="setActiveTab(id)"
+          v-show="!hide"
+          @click="tabClickHandler(id)"
         >
           <span class="TabName">{{ name | truncate( 21, '...') }}</span>
 
@@ -65,6 +66,14 @@
             @click="saveName(index)"
           >
             <span class="FontIcon name_checkBig"/>
+          </button>
+          <button
+            class="BtnIcon name_copy"
+            type="button"
+            title="Скопировать"
+            @click="copyTab(id, name)"
+          >
+            <span class="FontIcon name_copy"/>
           </button>
           <button
             class="BtnIcon type_delete"
@@ -112,6 +121,9 @@ export default {
       editMode: $root.editMode,
       visibleNavBar: $root.visibleNavBar,
       tabsSwitcherInstance: $root.tabsSwitcherInstance,
+      logSystem: $root.logSystem,
+      eventSystem: $root.eventSystem,
+      currentTab: '',
       tabName: '',
       scrollButton: false,
       scrollWidth: null,
@@ -166,6 +178,7 @@ export default {
         name: name ? name : tabId,
         isActive: false,
         editName: false,
+        hide: false,
       });
 
       this.tabsSwitcherInstance.addNewTab(tabId, this.tabsCollection);
@@ -174,12 +187,25 @@ export default {
       }
       return tabId;
     },
+
+    tabClickHandler(id) {
+      const clickedTab = this.getTab(id);
+
+      this.eventSystem.publishEvent('WorkspaceTabClicked', clickedTab);
+      this.logSystem.debug(`Clicked "${id}" workspace tab`);
+
+      if (id !== this.currentTab) this.setActiveTab(id);
+    },
+
     setActiveTab(tabId) {
       this.tabsCollection.forEach((tab) => {
         tab.isActive = tab.id === tabId;
       });
       this.tabsSwitcherInstance.activeTab(tabId);
+      this.currentTab = tabId;
+      this.logSystem.debug(`Activated "${tabId}" workspace tab`);
     },
+
     removeTab(tabId, index) {
       if (!tabId || this.tabsCollection.length === 1) {
         return;
@@ -193,19 +219,41 @@ export default {
       }
 
     },
+
     getTab(tabId) {
       if (!tabId) {
         return;
       }
       return this.tabsCollection.find((tab) => tab.id === tabId);
     },
+    toggleVisibleTabByName(tabName) {
+      let isTabHiddenActive = false;
+      this.tabsCollection.forEach((tab) => {
+        if (tab.name === tabName) {
+          tab.hide = !tab.hide;
+          isTabHiddenActive = tab.isActive;
+        }
+      });
+
+      if (isTabHiddenActive) {
+        for (let i = 0; i < this.tabsCollection.length; i++) {
+          const tab = this.tabsCollection[i];
+          if (!tab.hide) {
+            this.setActiveTab(tab.id);
+            break;
+          } 
+        }
+      }
+    },
     getIdNewTab() {
       return `wss-tab-${Math.ceil(Math.random() * 10000)}`;
     },
+
     changeName(index, name) {
       this.tabsCollection[index].editName = true;
       this.tabName = name;
     },
+
     saveName(index) {
       const checkTabName = this.tabsCollection.find((tab, i) => tab.name === this.tabName && i !== index);
       if (checkTabName) {
@@ -215,9 +263,11 @@ export default {
       this.tabsCollection[index].editName = false;
       this.tabsCollection[index].name = this.tabName;
     },
+
     resetError(index) {
       this.$refs.input[index].invalid = false;
     },
+
     scroll(toRight) {
       this.$refs.tabList.scrollLeft += toRight
           ? (this.$refs.tabPanel.clientWidth * 0.8)
@@ -225,6 +275,18 @@ export default {
       this.scrollWidth = this.$refs.tabList.scrollWidth;
       this.scrollLeft = this.$refs.tabList.scrollLeft;
       this.clientWidth = this.$refs.tabPanel.clientWidth;
+    },
+    copyTab(id, name) {
+      const tabId = this.getIdNewTab();
+      const numberName = this.tabsCollection.filter((item) => item.name === name || item.name?.includes(name)).length;
+      this.tabList = ({
+        id: tabId,
+        name: name+'-copy-'+numberName,
+        isActive: false,
+        editName: false,
+      });
+
+      this.tabsSwitcherInstance.copyTab(tabId, id, this.tabsCollection);
     },
   },
   filters: {