+++ DTCD-WorkspaceSystem/src/WorkspaceSystem.js
@@ -12,8 +12,8 @@ import {
 } from './../../DTCD-SDK/index';
 import { version } from './../package.json';
 
-import './styles/panel.css';
-import './styles/modal.css';
+import './styles/panel.scss';
+import './styles/modal.scss';
 import gridstackOptions from './utils/gridstackOptions';
 import TabsSwitcher from './TabsSwitcher';
 import utf8_to_base64 from './libs/utf8tobase64';
@@ -36,6 +36,8 @@ export class WorkspaceSystem extends SystemPlugin {
   #currentPath;
   #currentID;
   #column;
+  #typeInit;
+  #hiddenPanelPlugins;
 
   // ---- INTERNAL'S ----
   #activeGrid;
@@ -58,11 +60,21 @@ export class WorkspaceSystem extends SystemPlugin {
     };
   }
 
+  static INIT_TYPES = [
+    'TYPE-1',
+    'TYPE-2',
+  ];
+
   constructor(guid) {
     super();
     this.#guid = guid;
     this.#eventSystem = new EventSystemAdapter('0.4.0', guid);
-    this.#eventSystem.registerPluginInstance(this, ['WorkspaceCellClicked']);
+    this.#eventSystem.registerPluginInstance(this, [
+      'WorkspaceCellClicked',
+      'WorkspaceTabSelectedProgrammly',
+      'WorkspaceTabClicked',
+      'WorkspaceTitleLoaded',
+    ]);
     this.#interactionSystem = new InteractionSystemAdapter('0.4.0');
     this.#logSystem = new LogSystemAdapter('0.5.0', this.#guid, 'WorkspaceSystem');
     this.#styleSystem = new StyleSystemAdapter('0.5.0');
@@ -70,6 +82,8 @@ export class WorkspaceSystem extends SystemPlugin {
     this.#panels = [];
     this.#editMode = false;
     this.#modalInstance = null;
+    this.#typeInit = WorkspaceSystem.INIT_TYPES[0];
+    this.#hiddenPanelPlugins = [];
   }
 
   get currentWorkspaceTitle() {
@@ -86,6 +100,20 @@ export class WorkspaceSystem extends SystemPlugin {
 
   get currentWorkspaceColumn() {
     return this.#column;
+  }
+
+  get tabsCollection() {
+    return this.#tabsSwitcherInstance?.tabsCollection;
+  }
+
+  get typeInit() {
+    return this.#typeInit;
+  }
+
+  set typeInit(newValue) {
+    this.#typeInit = WorkspaceSystem.INIT_TYPES.includes(newValue)
+                    ? newValue
+                    : WorkspaceSystem.INIT_TYPES[0];
   }
 
   getFormSettings() {
@@ -190,6 +218,24 @@ export class WorkspaceSystem extends SystemPlugin {
             callback: this.#handleBorderColorChange.bind(this),
           },
         },
+        {
+          component: 'divider',
+        },
+        {
+          component: 'select',
+          propName: 'typeInit',
+          attrs: {
+            label: 'Варианты открытия рабочего стола',
+          },
+          handler: {
+            event: 'change',
+            callback: this.#handleTypeInitChange.bind(this),
+          },
+          options: [
+            { value: 'TYPE-1', label: 'Окрыватются сразу все вкладки' },
+            { value: 'TYPE-2', label: 'Открывается только активная' },
+          ],
+        }
       ],
     };
   }
@@ -236,6 +282,7 @@ export class WorkspaceSystem extends SystemPlugin {
     const data = {
       guid: this.#guid,
       interactionSystem: this.#interactionSystem,
+      logSystem: this.#logSystem,
       eventSystem: this.#eventSystem,
       plugin: this,
       tabsCollection: this.#tabsCollection,
@@ -251,13 +298,18 @@ export class WorkspaceSystem extends SystemPlugin {
 
     this.#tabsSwitcherInstance.htmlElement.appendChild(this.#vueComponent.$el);
     element.appendChild(this.#tabsSwitcherInstance.htmlElement);
-    
+
     this.#tabsSwitcherInstance.htmlElement.addEventListener('tab-active', this.#handleTabsSwitcherActive);
     this.#tabsSwitcherInstance.htmlElement.addEventListener('tab-delete', this.#handleTabsSwitcherDelete);
     this.#tabsSwitcherInstance.htmlElement.addEventListener('tab-add', this.#handleTabsSwitcherAdd);
+    this.#tabsSwitcherInstance.htmlElement.addEventListener('tab-copy', this.#handleTabsSwitcherCopy);
+    this.#eventSystem.subscribe(this.#guid, 'WorkspaceTitleLoaded', 'HeaderPanel_top', 'showTitle');
 
     const workspaceID = history.state.workspaceID;
-    this.setConfiguration(workspaceID);
+    this.setConfiguration(workspaceID)
+        .then(() => {
+          this.recoveryPluginStateFromUrl();
+        });
 
     return true;
   }
@@ -289,7 +341,15 @@ export class WorkspaceSystem extends SystemPlugin {
           tabId: panel?.position.tabId,
         };
         const undeletable = panel.undeletable;
-        plugins.push({ guid, meta, config, position, undeletable });
+        const toFixPanel = panel.toFixPanel;
+
+        plugins.push({ guid, meta, config, position, undeletable, toFixPanel });
+      });
+
+    // панели, которые не были созданы
+    this.#hiddenPanelPlugins
+      .forEach(panel => {
+        if (panel) plugins.push(panel);
       });
 
     const tabPanelsConfig = this.#vueComponent.getConfig;
@@ -299,6 +359,7 @@ export class WorkspaceSystem extends SystemPlugin {
       title: this.#currentTitle,
       column: this.#column,
       editMode: this.#editMode,
+      typeInit: this.typeInit,
       plugins,
       tabPanelsConfig,
       visibleTabNavBar: tabPanelsConfig.visibleNavBar,
@@ -335,6 +396,9 @@ export class WorkspaceSystem extends SystemPlugin {
     this.#currentTitle = config.title;
     this.#currentID = config.id;
     this.#currentPath = config.path;
+    this.typeInit = config.typeInit;
+
+    this.#eventSystem.publishEvent('WorkspaceTitleLoaded', this.#currentTitle);
 
     // ---- PLUGINS ----
 
@@ -343,19 +407,35 @@ export class WorkspaceSystem extends SystemPlugin {
     // ---- installing-plugins-from-config ----
     const GUIDMap = {};
     pluginsLoop: for (let plugin of config.plugins) {
-      let { meta, config, position = {}, guid } = plugin;
+      const {
+        meta,
+        config,
+        position = {},
+        guid,
+        toFixPanel,
+      } = plugin;
+
       switch (meta?.type) {
         case 'panel':
           let widget;
           if (typeof meta.name !== 'undefined') {
+
+            if (this.typeInit === 'TYPE-2') {
+              const isPanelOnActiveTab = position?.tabId === activeTabId;
+              if (!toFixPanel && !isPanelOnActiveTab) {
+                this.#hiddenPanelPlugins.push(plugin);
+                continue pluginsLoop;
+              }
+            }
+
             const pluginExists = this.getPlugin(meta.name, meta.version);
             if (pluginExists) {
               this.#logSystem.debug('Creating empty cell');
 
               // активирование таба нужно для корректной отрисовки визуализаций
-              if (position?.tabId && !position.isActive) {
-                this.#vueComponent.setActiveTab(position.tabId);
-              }
+              // if (position?.tabId && !position.isActive) {
+              //   this.#vueComponent.setActiveTab(position.tabId);
+              // }
 
               widget = this.createCell({
                 name: meta.name,
@@ -363,10 +443,12 @@ export class WorkspaceSystem extends SystemPlugin {
                 guid,
                 ...position,
                 autoPosition: false,
+                toFixPanel,
               });
             }
-            const plugin = this.#panels.find(panel => panel.widget === widget).instance;
-            const pluginGUID = this.getGUID(plugin);
+            
+            const pluginInstance = this.#panels.find(panel => panel.widget === widget).instance;
+            const pluginGUID = this.getGUID(pluginInstance);
             this.#logSystem.debug(`Mapping guid of ${meta.name} from ${guid} to ${pluginGUID}`);
             GUIDMap[guid] = pluginGUID;
           } else {
@@ -391,11 +473,15 @@ export class WorkspaceSystem extends SystemPlugin {
       }
 
       const instance = this.getInstance(GUIDMap[guid]);
-      if (instance && instance !== this && instance.setPluginConfig && config) await instance.setPluginConfig(config);
+      if (instance && instance !== this && instance.setPluginConfig && config) {
+        instance.setPluginConfig(config);
+      }
     }
 
     // активируем таб, который должен быть активным после открытия рабочего стола.
     this.#vueComponent.setActiveTab(activeTabId);
+    
+    this.#hideTabsPanel();
 
     // EVENT-SYSTEM-MAPPING
     if (eventSystemConfig.hasOwnProperty('subscriptions')) {
@@ -405,6 +491,11 @@ export class WorkspaceSystem extends SystemPlugin {
         action.guid = GUIDMap[action.guid];
       }
     }
+    this.#eventSystem.setPluginConfig(eventSystemConfig);
+
+    this.#panels.forEach((panel) => {
+      if (panel.toFixPanel) this.#createGridCellClones(panel.guid);
+    });
 
     // settings panel styles
     this.#panelStyles = {
@@ -414,18 +505,32 @@ export class WorkspaceSystem extends SystemPlugin {
     };
     this.#setPanelStyles();
 
-    await this.#eventSystem.setPluginConfig(eventSystemConfig);
     return true;
   }
 
   async downloadConfiguration(downloadPath) {
     const delimIndex = downloadPath.search(/:id=/);
-    const id = delimIndex !== -1 ? downloadPath.slice(delimIndex + 4) : downloadPath;
+    const id = delimIndex !== -1
+      ? downloadPath.split('?id=')[0].slice(delimIndex + 4)
+      : downloadPath.split('?id=')[0];
     const path = delimIndex !== -1 ? downloadPath.slice(0, delimIndex) : '';
+
+    const groups = await this.#interactionSystem.GETRequest('dtcd_utils/v1/user?photo_quality=low')
+    .then((response) => {
+      const groups = response.data.groups;
+      if (!groups?.length) return[];
+      return groups
+    });
+    const groupsForWorkSpaces = groups.filter(group => group.name.includes('workspace.')).map((item) => item.name.split('.')[1])
 
     this.#logSystem.debug(`Trying to download configuration with id:${id}`);
     const { data } = await this.#interactionSystem.GETRequest(`/dtcd_workspaces/v1/workspace/object/${path}?id=${id}`);
     this.#logSystem.debug(`Parsing configuration from response`);
+
+    if (groupsForWorkSpaces.includes(data.title)) {
+      Application.getSystem('RouteSystem', '0.3.0').navigate('/workspaces')
+    }
+
     const content = data.content;
     content['id'] = data.id;
     content['title'] = data.title;
@@ -535,7 +640,7 @@ export class WorkspaceSystem extends SystemPlugin {
     return maxID !== -Infinity ? maxID + 1 : 1;
   }
 
-  createCell({ name, version, guid = null, w = 6, h = 8, x = 0, y = 0, tabId, autoPosition = true }) {
+  createCell({ name, version, guid = null, w = 6, h = 8, x = 0, y = 0, tabId, autoPosition = true, toFixPanel }) {
     this.#logSystem.debug(
       `Adding panel-plugin widget with name:'${name}', version:${version}, w:${w},h:${h},x:${x},y:${y}, autoPosition:${autoPosition}`
     );
@@ -546,28 +651,14 @@ export class WorkspaceSystem extends SystemPlugin {
       guid = `${name}_${panelID}`;
     }
 
+    toFixPanel = Boolean(toFixPanel);
+
     let targetGrid = this.#gridCollection.get(tabId)?.gridInstance;
     targetGrid = targetGrid ? targetGrid : this.#activeGrid;
 
-    // TODO: Replace on WEB-COMPONENT with style!
-    const widget = targetGrid.addWidget(
-      `
-      <div class="grid-stack-item">
-        <div class="grid-stack-item-content">
-          <div class="handle-drag-of-panel gridstack-panel-header" style="display:${this.#editMode ? 'flex' : 'none'}">
-            <div id="closePanelBtn-${guid}" class="close-panel-button">
-              <span class="FontIcon name_closeBig size_lg"></span>
-            </div>
-            <span class="drag-panel-button FontIcon name_move size_lg"></span>
-          </div>
-          <div class="gridstack-content-container${this.#editMode ? ' gridstack-panel-overlay' : ''}">
-            <div id="panel-${guid}">
-            </div>
-          </div>
-        </div>
-      </div>
-      `,
-      { x, y, w, h, autoPosition, id: guid }
+    const widget = this.#createWidget(
+      targetGrid,
+      { x, y, w, h, autoPosition, guid, toFixPanel }
     );
 
     const panelInstance = this.installPanel({
@@ -576,12 +667,6 @@ export class WorkspaceSystem extends SystemPlugin {
       version,
       selector: `#panel-${guid}`,
     });
-
-    widget.addEventListener('click', () => {
-      if (!this.#editMode) this.#eventSystem.publishEvent('WorkspaceCellClicked', { guid });
-    });
-
-    document.getElementById(`closePanelBtn-${guid}`).addEventListener('click', this.deleteCell.bind(this, guid));
 
     const meta = panelInstance.constructor.getRegistrationMeta();
 
@@ -593,7 +678,14 @@ export class WorkspaceSystem extends SystemPlugin {
       instance: panelInstance,
       guid,
       meta,
+      toFixPanel,
     });
+
+    // отключил этот код, так как при инициализации рабочего стола
+    // панели в сетках устанавливаются не так, как их сохранили.
+    // if (toFixPanel) {
+    //   this.#createGridCellClones(guid);
+    // }
 
     return widget;
   }
@@ -606,17 +698,44 @@ export class WorkspaceSystem extends SystemPlugin {
       return;
     }
 
-    this.#panels.splice(
-      this.#panels.findIndex(plg => plg.guid === guid),
-      1
-    );
-
+    if (panel.toFixPanel) this.#deleteGridCellClones(panel.guid);
     const targetGrid = this.#gridCollection.get(panel.position.tabId).gridInstance;
     targetGrid.removeWidget(panel.widget);
 
     this.uninstallPluginByGUID(guid);
 
+    this.#panels.splice(
+      this.#panels.findIndex(plg => plg.guid === guid),
+      1
+    );
+
     this.#logSystem.info(`Deleted cell from workspace with guid: ${guid}`);
+  }
+
+  toggleFixPanel(guid) {
+    this.#logSystem.debug(`Start toggle fixation of cell on workspace with guid: ${guid}`);
+
+    const panel = this.#panels.find(panel => panel.guid === guid);
+    if (!panel) {
+      this.#logSystem.debug(`No cell element found on workspace with given guid: ${guid}`);
+      return;
+    }
+
+    panel.toFixPanel = !Boolean(panel.toFixPanel);
+    const targetGrid = this.#gridCollection.get(panel.position.tabId).gridInstance;
+    targetGrid.update(
+      panel.widget,
+      {
+        noMove: panel.toFixPanel,
+        noResize: panel.toFixPanel,
+        locked: panel.toFixPanel,
+      }
+    );
+
+    if (panel.toFixPanel) this.#createGridCellClones(guid);
+    else this.#deleteGridCellClones(guid);
+
+    this.#logSystem.info(`End toggle fixation of cell on workspace with guid: ${guid}`);
   }
 
   compactAllPanels() {
@@ -629,28 +748,10 @@ export class WorkspaceSystem extends SystemPlugin {
   changeMode() {
     this.#editMode = !this.#editMode;
 
-    const panelBorder = document.querySelectorAll('.grid-stack-item-content');
-    panelBorder.forEach(content => {
-      if (this.#editMode) {
-        if (this.#panelStyles['border-width']) {
-          content.style.border = `${this.#panelStyles['border-width']} solid var(--button_primary)`;
-        } else {
-          content.style.border = '2px solid var(--button_primary)';
-        }
-      } else {
-        content.style.border = '';
-      }
-    });
-
-    const panelHeaders = document.querySelectorAll('.gridstack-panel-header');
-    panelHeaders.forEach(header => {
-      header.style.display = this.#editMode ? '' : 'none';
-    });
-
-    const overlayClass = 'gridstack-panel-overlay';
-    const panelContents = document.querySelectorAll('.gridstack-content-container');
-    panelContents.forEach(content => {
-      this.#editMode ? content.classList.add(overlayClass) : content.classList.remove(overlayClass);
+    const panelBorder = document.querySelectorAll('.grid-stack-item');
+    panelBorder.forEach((gridCell) => {
+      if (this.#editMode) gridCell.classList.add('grid-stack-item_editing');
+      else gridCell.classList.remove('grid-stack-item_editing');
     });
 
     const margin = this.#editMode ? '2px' : '0px';
@@ -705,7 +806,7 @@ export class WorkspaceSystem extends SystemPlugin {
           content: this.getPluginConfig(),
         },
       ]);
-  
+
       this.#notificationSystem && this.#notificationSystem.create(
         'Готово!',
         'Настройки рабочего стола сохранены.',
@@ -814,6 +915,131 @@ export class WorkspaceSystem extends SystemPlugin {
     }
   }
 
+  /**
+   * Send all plugin states on workspace.
+   * @returns {Promise<string>} Promise containing a link of the dashboard with stateID
+   */
+  async getURLDashboardState() {
+    this.#logSystem.debug('Start creating URL with plugin states.');
+
+    try {
+      const state = this.#collectStatesFromPlugins();
+      const response = await this.#interactionSystem.POSTRequest(
+        'dtcd_storage_system_backend/v1/state',
+        {
+          applicationName: 'DataCAD',
+          workspaceID: this.#currentID,
+          state,
+        }
+      );
+      response.then((result) => {
+        const {
+          stateID,
+        } = JSON.parse(result);
+
+        if (stateID) {
+          const {
+            origin,
+            pathname,
+          } = window.location;
+          return origin + pathname + `?stateID="${stateID}"`;
+        }
+      });
+    } catch (error) {
+      this.#logSystem.error('Error creating URL with plugin states: ' + error.message);
+      throw error;
+    }
+  }
+
+  /**
+   * Recovery dashboard state from URL.
+   * @param {String} [url] URL with search parameter 'stateID'
+   */
+  recoveryPluginStateFromUrl(url) {
+    this.#logSystem.debug('Start dashboard state recovery from URL.');
+
+    try {
+      const urlSearchParams = new URL(
+        typeof url === 'string' ? url : window.location.href
+      ).searchParams;
+      const stateID = urlSearchParams.get('stateID');
+      if (!stateID) return;
+
+      const response = this.#interactionSystem.POSTRequest(
+        'dtcd_storage_system_backend/v1/state',
+        {
+          applicationName: 'DataCAD',
+          stateID,
+        },
+      );
+      response.then((result) => {
+        const dashboardState = JSON.parse(result).state;
+        for (const key in dashboardState) {
+          if (!Object.hasOwnProperty.call(dashboardState, key)) continue;
+
+          try {
+            this.#logSystem.debug(`Start plugin state recovery of '${key}' from URL.`);
+            if (typeof Application.autocomplete[key]?.setState === 'function') {
+              Application.autocomplete[key].setState(dashboardState[key]);
+            }
+          } catch (error) {
+            this.#logSystem.error(`Error plugin state recovery of '${key}' from URL.`);
+            console.error(error);
+          }
+        }
+      }).catch((error) => {
+        throw error;
+      });
+    } catch (error) {
+      this.#logSystem.error('Error recovery dashboard state from URL: ' + error.message);
+      this.#notificationSystem && this.#notificationSystem.create(
+        'Error.',
+        'Ошибка восстановления состояния рабочего стола из URL.',
+        {
+          floatMode: true,
+          floatTime: 5,
+          type: 'error',
+        }
+      );
+      throw error;
+    }
+
+    this.#logSystem.info('Ended dashboard state recovery from URL.');
+    this.#notificationSystem && this.#notificationSystem.create(
+      'Выполнено.',
+      'Данные рабочего стола успешно восстановлены из URL.',
+      {
+        floatMode: true,
+        floatTime: 5,
+        type: 'success',
+      }
+    );
+  }
+
+  #hideTabsPanel() {
+    this.#interactionSystem.GETRequest('dtcd_utils/v1/user?photo_quality=low')
+      .then((response) => {
+        const groups = response.data.groups;
+        if (!groups?.length) return;
+        
+        for (let i = 0; i < groups.length; i++) {
+          this.#vueComponent.toggleVisibleTabByName(groups[i].name);
+        }
+      });
+  }
+
+  #collectStatesFromPlugins() {
+    const pluginsState = {};
+
+    this.#panels.forEach((panel) => {
+      if (typeof panel.instance?.getState === 'function') {
+        pluginsState[panel.guid] = panel.instance.getState();
+      }
+    });
+
+    return pluginsState;
+  }
+
   #createTabsSwitcher() {
     this.#gridCollection = new Map();
     this.#tabsCollection = [];
@@ -853,9 +1079,79 @@ export class WorkspaceSystem extends SystemPlugin {
       isActive: false,
       gridInstance: newGrid,
     });
+
+    this.#panels.forEach((panel) => {
+      if (panel.toFixPanel) this.#createGridCellClones(panel.guid);
+    });
   };
 
-  #handleTabsSwitcherActive = event => {
+  #handleTabsSwitcherCopy = event => {
+    const tabId = event.detail?.tabId;
+    const collection = event.detail?.collection;
+    const id = event.detail?.id;
+    const plugins = this.getPluginConfig().plugins;
+
+    collection.forEach((tab) => {
+      // Формирования списка плагинов
+      const targetPlugins =  plugins.reduce((acc, plugin) => {
+        if (plugin.meta?.type === 'panel') {
+          if (plugin?.position?.tabId === tab?.id && plugin?.position?.tabId === id) {
+
+            acc.push(plugin)
+          }
+        }
+        return acc
+      },[]);
+
+      // инифиализация плагинов на новой вкладке
+      const pluginsGuid = targetPlugins.reduce((acc, plugin) => {
+        const {name, version} = plugin.meta
+        const {h, w, x, y} = plugin.position
+        const autoPosition = false
+        const toFixPanel = false
+        const widget = this.createCell({ name, version, guid:null, w, h, x, y, tabId, autoPosition, toFixPanel });
+        const currentPlugin = Application.autocomplete[widget.getAttribute('gs-id')];
+        if (plugin.config) {
+          currentPlugin.setPluginConfig(plugin.config);
+        }
+        acc.push({
+          originPlugin: plugin.guid,
+          targetPlugin: widget.getAttribute('gs-id'),
+        })
+        return acc
+      }, []);
+
+      pluginsGuid.forEach((guid) => {
+        const subscriptions = this.#eventSystem.subscriptions.filter((item) => item.event.guid === guid.originPlugin && item.subscriptionName !== undefined)
+         if (subscriptions.length > 0) {
+           subscriptions.forEach((event) => {
+             let eventGuid, actionGuid = null;
+             if (event.subscriptionName) {
+               eventGuid = pluginsGuid.find((cellGuid) => cellGuid.originPlugin === event.event.guid);
+               if (event.action.guid) {
+                 actionGuid = pluginsGuid.find((cellGuid) => cellGuid.originPlugin === event.action.guid);
+               } else {
+                 actionGuid = event.action.guid;
+               }
+               if (!actionGuid?.targetPlugin) {
+                 this.#eventSystem.registerCustomAction(event.action.name+'-'+tabId, event.action.callback)
+               }
+               this.#eventSystem.subscribe({
+                 eventGUID: eventGuid.targetPlugin,
+                 eventName: event.event.name,
+                 actionGUID: actionGuid?.targetPlugin ? actionGuid?.targetPlugin : 'Пользовательское событие',
+                 actionName: actionGuid?.targetPlugin ? event.action.name : event.action.name+'-'+tabId,
+                 eventArgs: event.action.args,
+                 subscriptionName: event.subscriptionName+'-'+tabId,
+               });
+             }
+           })
+         }
+      });
+    });
+  };
+
+  #handleTabsSwitcherActive = (event) => {
     if (!this.#gridCollection) return;
 
     const activeTabId = event.detail.tabId;
@@ -872,24 +1168,77 @@ export class WorkspaceSystem extends SystemPlugin {
 
     this.#setTabIdUrlParam(activeTabId);
 
+    if (this.typeInit === 'TYPE-2') {
+      this.#createPanelsInActiveTab(activeTabId);
+    }
+
     this.#panels.forEach((panel) => {
+      if (panel.toFixPanel) {
+        this.#changeFixedPanelPosition(panel);
+      }
       if (typeof panel.instance.setVisible === 'function') {
         panel.instance.setVisible(activeTabId === panel?.position.tabId)
       }
     })
   };
 
+  async #createPanelsInActiveTab(activeTabId) {
+    for (let i = 0; i < this.#hiddenPanelPlugins.length; i++) {
+      const plugin = this.#hiddenPanelPlugins[i];
+      if (!plugin) continue;
+
+      const {
+        meta,
+        config,
+        position = {},
+        guid,
+        toFixPanel,
+      } = plugin;
+
+      if (position.tabId !== activeTabId) continue;
+
+      switch (meta?.type) {
+        case 'panel':
+          let widget;
+          if (typeof meta.name !== 'undefined') {
+            const pluginExists = this.getPlugin(meta.name, meta.version);
+            if (pluginExists) {
+              widget = this.createCell({
+                name: meta.name,
+                version: meta.version,
+                guid,
+                ...position,
+                autoPosition: false,
+                toFixPanel,
+              });
+            }
+          }
+          break;
+      }
+
+      const instance = this.getInstance(guid);
+      if (instance && instance !== this && instance.setPluginConfig && config) {
+        instance.setPluginConfig(config);
+      }
+
+      this.#hiddenPanelPlugins[i] = null;
+    }
+  }
+
   #setTabIdUrlParam(tabId) {
     if (!tabId) return;
 
-    const urlSearchParams = new URLSearchParams(window.location.search);
-    urlSearchParams.set('ws-tab-id', tabId);
-
-    Application.getSystem('RouteSystem', '0.3.0').navigate(
-      `${window.location.pathname}?${urlSearchParams.toString()}`,
-      true,
-      {workspaceID: this.currentWorkspaceID},
-    )
+    if (this.currentWorkspaceID) {
+      const urlSearchParams = new URLSearchParams(window.location.search);
+      urlSearchParams.set('ws-tab-id', tabId);
+      const currentWorkspaceSlug = window.location.pathname
+      .replace('/workspaces/','').split('?')[0]
+      Application.getSystem('RouteSystem', '0.3.0').navigate(
+        `${window.location.pathname}?${urlSearchParams.toString()}`,
+        true,
+        {workspaceID: currentWorkspaceSlug},
+      )
+    }
   }
 
   #getTabIdUrlParam() {
@@ -897,7 +1246,7 @@ export class WorkspaceSystem extends SystemPlugin {
     return urlSearchParams.get('ws-tab-id');
   }
 
-  #handleTabsSwitcherDelete = event => {
+  #handleTabsSwitcherDelete = (event) => {
     if (!this.#gridCollection) return;
 
     const deletingTabId = event.detail.tabId;
@@ -912,7 +1261,10 @@ export class WorkspaceSystem extends SystemPlugin {
     this.#panels = this.#panels.filter(plugin => {
       const { widget, guid, position } = plugin;
       if (position.tabId === deletingTabId) {
-        widget && deletingGrid.removeWidget(widget);
+        if (widget) {
+          this.#deleteGridCellClones(guid);
+          deletingGrid.removeWidget(widget);
+        }
         this.uninstallPluginByGUID(guid);
         return false;
       } else {
@@ -995,5 +1347,176 @@ export class WorkspaceSystem extends SystemPlugin {
     borderStyles += '}';
 
     this.#wssStyleTag.textContent = borderStyles;
+  }
+
+  #handleTypeInitChange = (event) => {
+    const { value } = event.target;
+    this.typeInit = value;
+  }
+
+  #createGridCellClones(guid) {
+    this.#logSystem.debug(`Start of creation grid cell clones for panel ${guid}.`);
+
+    const panel = this.#panels.find(panel => panel.guid === guid);
+    if (!panel) {
+      this.#logSystem.debug(`No cell element found on workspace with given guid: ${guid}`);
+      return;
+    }
+    const tabId = panel.position.tabId;
+    const { h, w, x, y } = panel?.widget.gridstackNode;
+
+    this.#gridCollection.forEach((gridData, key) => {
+      if (key === tabId) return;
+
+      let isExistGridCell = false;
+      gridData.gridInstance.getGridItems().forEach((gridCell) => {
+        // find doubles grid items
+        if (gridCell.getAttribute('gs-id') === guid) {
+          isExistGridCell = true;
+          return;
+        }
+      });
+
+      if (!isExistGridCell) {
+        this.#createWidget(
+          gridData.gridInstance,
+          { x, y, w, h, autoPosition: false, guid, toFixPanel: true, empty: true, }
+        );
+      }
+    });
+
+    this.#logSystem.info(`End of creation grid cell clones for panel ${guid}.`);
+  }
+
+  #deleteGridCellClones(guid) {
+    this.#logSystem.debug(`Start of deleting grid cell clones for panel ${guid}.`);
+
+    const panel = this.#panels.find(panel => panel.guid === guid);
+    if (!panel) {
+      this.#logSystem.debug(`No cell element found on workspace with given guid: ${guid}`);
+      return;
+    }
+
+    this.#gridCollection.forEach((gridData, key) => {
+      gridData.gridInstance.getGridItems().forEach((gridCell) => {
+        if (gridCell.getAttribute('gs-id') === guid) {
+          if (gridCell.hasAttribute('data-empty-item')) {
+            gridData.gridInstance.removeWidget(gridCell);
+            return;
+          }
+        }
+      });
+    });
+
+    this.#logSystem.info(`End of deleting grid cell clones for panel ${guid}.`);
+  }
+
+  #createWidget(targetGrid, gridItemOptions) {
+    this.#logSystem.debug(`Start of creation grid item cell.`);
+
+    const {
+      guid = null,
+      id = guid,
+      w = 6,
+      h = 8,
+      x = 0,
+      y = 0,
+      autoPosition = true,
+      toFixPanel,
+      empty,
+    } = gridItemOptions;
+
+    const widget = targetGrid.addWidget(
+      `
+      <div
+        class="grid-stack-item${this.#editMode ? ' grid-stack-item_editing' : ''}"
+        ${empty ? ' data-empty-item' : ''}
+      >
+        <div class="grid-stack-item-content">
+          <div class="handle-drag-of-panel gridstack-panel-header">
+            <button
+              class="fix-panel-button"
+              type="button"
+              title="Зафиксировать панель"
+            >
+              <span class="FontIcon name_location size_lg"></span>
+            </button>
+            <button
+              class="close-panel-button"
+              type="button"
+              title="Удалить панель"
+            >
+              <span class="FontIcon name_closeBig size_lg"></span>
+            </button>
+            <button
+              class="drag-panel-button"
+              type="button"
+              title="Переместить панель"
+            >
+              <span class="FontIcon name_move size_lg"></span>
+            </button>
+          </div>
+          <div class="gridstack-content-container">
+            ${empty ? '' : `<div id="panel-${guid}"></div>`}
+          </div>
+        </div>
+      </div>
+      `,
+      {
+        x, y, w, h, autoPosition, id,
+        locked: toFixPanel, noMove: toFixPanel, noResize: toFixPanel,
+      }
+    );
+
+    widget.addEventListener('click', () => {
+      if (!this.#editMode) this.#eventSystem.publishEvent('WorkspaceCellClicked', { guid });
+    });
+
+    widget.querySelector('.close-panel-button')
+          .addEventListener('click', this.deleteCell.bind(this, guid));
+    widget.querySelector('.fix-panel-button')
+          .addEventListener('click', this.toggleFixPanel.bind(this, guid));
+
+    this.#logSystem.info(`End of creation grid item cell.`);
+    return widget;
+  }
+
+  #changeFixedPanelPosition(panel) {
+    this.#logSystem.debug(`Start to replace fixed panel (guid: ${panel.guid}).`);
+
+    const htmlOfPanelInstance = panel.widget.querySelector('.gridstack-content-container > *');
+    if (!htmlOfPanelInstance || !htmlOfPanelInstance instanceof HTMLElement) {
+      this.#logSystem.debug('HTML element of panel not found.');
+      return;
+    }
+
+    panel.widget.setAttribute('data-empty-item', '');
+
+    this.#activeGrid.getGridItems().forEach((gridCell) => {
+      if (gridCell.getAttribute('gs-id') === panel.guid) {
+        gridCell.querySelector('.gridstack-content-container').append(htmlOfPanelInstance);
+        gridCell.removeAttribute('data-empty-item');
+        panel.position.tabId = this.#getGridIdByObject(this.#activeGrid);
+        panel.widget = gridCell;
+        return;
+      }
+    });
+
+    this.#logSystem.info(`End to replace fixed panel (guid: ${panel.guid}).`);
+  }
+
+  setActiveTab(tabID) {
+    this.#logSystem.debug(`Trying to select tab: ${tabID}`);
+
+    const tab = this.tabsCollection.find(tab => tabID === tab.id);
+
+    if (!tab) {
+      return this.#logSystem.debug(`Tab with ID ${tabID} not found`);;
+    }
+
+    this.#vueComponent.setActiveTab(tabID);
+
+    this.#logSystem.debug(`Tab selected: ${tabID}`);
+    this.#eventSystem.publishEvent('WorkspaceTabSelectedProgrammly', tab);
   }
 }